sched/fair: Make sd_overutilized if cpu is overutilized
update_overutilized_status() is used to set the overutilized flag when a
task gets queued up on a cpu or during the tick. The primary purpose is
to check if cpu is running close to its capacity and set the
overutilized flag to allow load balancing on that domain.

However, change f5438882fb2ccb64b55e169df1c5e76f5b2b7e87 ("sched: fair:
mark parent domain overutilized if !SD_LOAD_BALANCE"), accidentally
removed the check to see if the cpu is overutilized or not.
This leads to setting the overutilized flag in the sd domain when it
might not necessarily be overutilized, causing load balancer to
unnecesarily run on that sd domain.

Change-Id: Iaef3a660b7af6a073158dfd530973b974b28822d
Signed-off-by: Runmin Wang <runminw@codeaurora.org>
Signed-off-by: nysascape <jago@nysascape.digital>
Signed-off-by: Dušan Uverić <dusan.uveric9@gmail.com>
Signed-off-by: Reinazhard <muh.alfarozy@gmail.com>